% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_relative_performance.R
\name{bootstrap_relative_performance}
\alias{bootstrap_relative_performance}
\title{Bootstrap the performance statistic as explained in Rodriguez & Spirling 2020}
\usage{
bootstrap_relative_performance(
  tt_tibble,
  x = NULL,
  models = NULL,
  overlap_prob = NULL,
  num_iters = 100,
  seed = 1984L
)
}
\arguments{
\item{x}{the cue for which performance statistic is to be bootstrapped}

\item{overlap_prob}{numeric between 0-1, output of \code{compute_overlap_prob} for the corresponding models}

\item{num_iters}{number of bootstraps}

\item{seed}{seed for replication purposes}

\item{tt_data}{processed triad task app output}
}
\value{

}
\description{
Bootstrap the performance statistic as explained in Rodriguez & Spirling 2020
}
