% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_overlap_prob.R
\name{compute_overlap_prob}
\alias{compute_overlap_prob}
\title{Compute the probability that the same word is drawn by chance given two lists of nearest neighbors}
\usage{
compute_overlap_prob(nns1, nns2)
}
\arguments{
\item{nns1}{character vector of nearest neighbors 1}

\item{nns2}{character vector of nearest neighbors 2}
}
\value{
the probability (numeric) that the same word is drawn by chance
}
\description{
Compute the probability that the same word is drawn by chance given two lists of nearest neighbors
}
