% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_relative_performance.R
\name{compute_relative_performance}
\alias{compute_relative_performance}
\title{Compute the performance statistic as explained in Rodriguez & Spirling 2020}
\usage{
compute_relative_performance(tt_data, models, overlap_prob = NULL)
}
\arguments{
\item{tt_data}{processed triad task app output}

\item{overlap_prob}{numeric between 0-1, output of \code{compute_overlap_prob} for the corresponding models}

\item{nns2}{character vector specifying pair of models to compare (must match labels used in tt_data)}
}
\value{

}
\description{
Compute the performance statistic as explained in Rodriguez & Spirling 2020
}
